requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
gunicorn==22.0.0
Sastrawi==1.0.1
gensim==4.3.2
